distract_the_trainers (#6)

+++ src/year_2022/level_4/distract_the_trainers/solution.py
@@ -9,7 +9,7 @@ def is_power_of_two(x):
     return x & (x - 1) == 0
 
 def is_infinite_pair(x, y):
-    return not is_power_of_two((x + y) / gcd(x, y))
+    return not is_power_of_two(int((x + y) / gcd(x, y)))
 
 def create_graph(items):
     graph = { i: [] for i in range(len(items)) }
